Aula 92

+++ src/main/java/br/com/erudio/dto/Exchange.java
@@ -12,14 +12,14 @@ public class Exchange implements Serializable {
     private Long id;
     private String from;
     private String to;
-    private BigDecimal conversionFactor;
-    private BigDecimal convertedValue;
+    private Double conversionFactor;
+    private Double convertedValue;
     private String environment;
 
     public Exchange() {
     }
 
-    public Exchange(Long id, String from, String to, BigDecimal conversionFactor, BigDecimal convertedValue, String environment) {
+    public Exchange(Long id, String from, String to, Double conversionFactor, Double convertedValue, String environment) {
         this.id = id;
         this.from = from;
         this.to = to;
@@ -44,11 +44,11 @@ public class Exchange implements Serializable {
         this.to = to;
     }
 
-    public BigDecimal getConversionFactor() {
+    public Double getConversionFactor() {
         return conversionFactor;
     }
 
-    public void setConversionFactor(BigDecimal conversionFactor) {
+    public void setConversionFactor(Double conversionFactor) {
         this.conversionFactor = conversionFactor;
     }
 
@@ -60,11 +60,11 @@ public class Exchange implements Serializable {
         this.from = from;
     }
 
-    public BigDecimal getConvertedValue() {
+    public Double getConvertedValue() {
         return convertedValue;
     }
 
-    public void setConvertedValue(BigDecimal convertedValue) {
+    public void setConvertedValue(Double convertedValue) {
         this.convertedValue = convertedValue;
     }
